vijay gen ai project final/requirements.txt
numpy
pandas
langchain
openai
google-generativeai
transformers
datasets
accelerate
rank_bm25
sentence_transformers 
sklearn.metrics.pairwise 
ast
faiss
chroma
genai
re
json
















